LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/canvas/element/manual/drawing-images-to-the-canvas/drawimage_crossorigin.sub-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.

PASS Can get pixels of canvas with same origin image drawn
PASS Can get pixels of canvas with CORS enabled cross origin image drawn
PASS Can't get pixels of canvas with CORS enabled cross origin image drawn from non-CORS element
PASS Can't get pixels of canvas with non-CORS enabled cross origin image drawn
